Make code able to process other paths:
- Allow reading only M1-5 and N1-3 ROOT files data from TB2017, generated by Jordi
which are located at /eos/user/d/duarte/alibavas_data_root/. Extaction of
sensor name and run number from the path has changed. So:

- Creation of a variable ("selection") which prepares the code for reading
M1-5 and N1-3 TB2017 ROOT files generated by Jordi using fortythieves.

- Type of measurement (TB2017) cannot be read from the path, so it is
selected directly in the code when it detects the type of directory
which is reading.